Explicitly convert duration to double.
Since we're only expecting values in the order of 5, this should be
fine. This fixes the warning C4244 on MSVC Version 19.28.29914.

Machine code produced is the same before and after.

+++ PrimeCPP/PrimeCPP.cpp
@@ -127,7 +127,7 @@ int main()
         passes++;
         if (duration_cast<seconds>(steady_clock::now() - tStart).count() >= 5)
         {
-            sieve.printResults(false, duration_cast<microseconds>(steady_clock::now() - tStart).count() / 1000000, passes);
+            sieve.printResults(false, static_cast<double>(duration_cast<microseconds>(steady_clock::now() - tStart).count() / 1000000), passes);
             break;
         }
     }
